requirements.txt
webdriver-manager>=3.8.5
selenium>=4.8.0

# Essential testing utilities
pytest-html>=3.2.0
pytest-timeout>=2.1.0
pytest-xdist>=3.8.0

# Reporting
allure-pytest>=2.13.0
allure-python-commons>=2.13.0

# Additional dependencies for enhanced functionality
python-dotenv>=1.0.0
